refactor: intervalTime 관련 로직 customHook으로 추상화
